[BE] 애플 로그인 revoke 요청이 실패하는 오류 수정

+++ src/main/java/site/dogether/auth/service/AuthService.java
@@ -38,7 +38,9 @@ public class AuthService {
 
     @Transactional
     public void withdraw(final Long memberId, final WithdrawRequest request) {
-        appleOAuthProvider.revoke(request.authorizationCode());
-        memberService.delete(memberId);
+        boolean isRevoked = appleOAuthProvider.revoke(request.authorizationCode());
+        if (isRevoked) {
+            memberService.delete(memberId);
+        }
     }
 }
